added: schedule endpoints
