+++ src/configs/config-api.ts
@@ -1,3 +1,3 @@
-export const config = {
+export const configApi = {
   baseScheduleAPI: 'http://212.193.62.200:3007/api/',
 };
